Can't remove qaqc | calibrated if they don't exist
When adding a new site, qaqc and calibrated directories are not automatically created. But because they don't exist, the reprocess flag is automatically set to True, proc_init then tries to delete them because the reprocess flag is true, but they don't exist yet.

+++ src/proc_init.r
@@ -21,7 +21,11 @@ proc_init <- function() {
   if (site_config$reprocess) {
     message('Reprocessing data archive for: ', site, '/', instrument)
     for (path in file.path(wd, c('qaqc', 'calibrated'))) {
-      system(paste('rm -r', path))
+      if (file.exists(path)) {
+            system(paste('rm -r', path))
+      } else {
+        message('Path does not exist:', path)
+      }
     }
   }
